Split project into multiple packages for #11
