+++ client/package.yaml
@@ -1,22 +1,13 @@
-# This YAML file describes your package. Stack will automatically generate a
-# Cabal file when you run `stack build`. See the hpack website for help with
-# this file: <https://github.com/sol/hpack>.
-name: water-wars
+name: water-wars-client
 version: '0.0.0'
 github: "FailWhaleBrigade/water-wars"
 license: MIT
 author: "Fendor, VeryMilkyJoe, Power-Fungus"
 maintainer: "Fendor, VeryMilkyJoe, Power-Fungus"
-# synopsis: A new Haskeleton package.
-# description: water-wars is a new Haskeleton package.
-# category: Other
+synopsis: Water Wars client library
 
 extra-source-files:
-- CHANGELOG.md
-- LICENSE.md
 - package.yaml
-- README.md
-- stack.yaml
 
 ghc-options:
 - -Wall
@@ -41,7 +32,8 @@ library:
   - cereal
   - monad-logger
   - optparse-applicative
-  source-dirs: library
+  - water-wars
+  source-dirs: src
   default-extensions:
   - DeriveGeneric
   - NamedFieldPuns
@@ -53,42 +45,27 @@ library:
   - FlexibleContexts
   - MonoLocalBinds
 
-benchmarks:
-  water-wars-benchmarks:
-    source-dirs: benchmark
+executables:
+  water-wars-client:
+    source-dirs: app
     main: Main.hs
     dependencies:
     - base
     - water-wars
-    - criterion
-    ghc-options:
-    - -rtsopts
-    - -threaded
-    - -with-rtsopts=-N
-    default-extensions:
-    - NamedFieldPuns
-    - OverloadedStrings
-    - NoImplicitPrelude
-    - ScopedTypeVariables
-    - RecordWildCards
-
-tests:
-  water-wars-test-suite:
-    source-dirs: test-suite
-    main: Main.hs
-    dependencies:
-    - base
-    - water-wars
+    - water-wars-client
     - classy-prelude
-    - tasty
-    - tasty-hspec
-    - tasty-hunit
-    - hspec
-    - HUnit
-    - array
+    - ekg
+    - transformers
+    - mtl
+    - gloss  
+    - stm
+    - proteaaudio
+    - monad-logger
+    - optparse-applicative
     ghc-options:
     - -rtsopts
     - -threaded
+    - -O2
     - -with-rtsopts=-N
     default-extensions:
     - NamedFieldPuns
@@ -96,3 +73,5 @@ tests:
     - NoImplicitPrelude
     - ScopedTypeVariables
     - RecordWildCards
+    - FlexibleContexts
+
