+++ client/src/WaterWars/Client/Codec/Resource.hs
