--- test-suite/WaterWars/Server/GameNgTest.hs
@@ -1,1 +1,0 @@
-module WaterWars.Server.GameNgTest where
